'RAVE' class for base repository

The class is for creating child classes, to instantiate the class, please use prepare_subject_bare0 to create base repository.

See also

Super class

ravepipeline::RAVESerializable -> RAVESubjectRepository

Public fields

repository_id

repository identifier, typically generated with random string

quiet

whether to suppress update warning messages, when requested electrodes are not fully processed or excluded

Active bindings

needs_update

write-only attribute when subject needs to be reloaded from the disk and reference table needs to be updated, use repo$needs_update <- TRUE

project

project instance, see RAVEProject

subject

subject instance, see RAVESubject

electrode_list

integer vector of electrodes included

electrode_table

the entire electrode table

electrode_signal_types

more accurate name should be "channel" signal types: currently returns 'LFP', 'Auxiliary', or 'Spike', for each channel

electrode_instances

electrode channel instance helpers for loading electrode data

reference_name

name of reference table

reference_table

reference table

references_list

a vector of reference channel names, used together with reference_instances

reference_instances

instances of reference channels, for referencing on the fly, used for electrode_instances

digest_key

a list of repository data used to generate repository signature

signature

signature of the repository, two repositories might share the same signature if their contents are the same (even with different identifiers); generated from digest_key

Methods

Public methods

RAVESubjectBaseRepository$@marshal()

RAVESubjectBaseRepository$@unmarshal()

RAVESubjectBaseRepository$new()

RAVESubjectBaseRepository$clone()

Inherited methods

ravepipeline::RAVESerializable$@compare()

Method @marshal()

Internal method

Usage

RAVESubjectBaseRepository$@marshal(...)

Arguments

...

internal arguments

Method @unmarshal()

Internal method

Usage

RAVESubjectBaseRepository$@unmarshal(object, ...)

Arguments

object, ...

internal arguments

Method new()

constructor

Usage

RAVESubjectBaseRepository$new(
  subject,
  electrodes = NULL,
  reference_name = NULL,
  ...,
  quiet = TRUE,
  repository_id = NULL,
  strict = TRUE,
  .class = NULL
)

Arguments

subject

'RAVE' subject

electrodes

string or integers indicating electrodes to load

reference_name

name of the reference table

...

reserved, currently ignored

quiet

see field quiet

repository_id

see field repository_id

strict

whether the mode should be strict; default is true and errors out when subject is missing

.class

internally used, do not set, even if you know what this is

Method clone()

The objects of this class are cloneable with this method.

Usage

RAVESubjectBaseRepository$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.