Create brain atlases from template

Reverse-transform from template atlases via volume mapping

Usage

generate_atlases_from_template(
  subject,
  atlas_folders,
  template_name = "mni_icbm152_nlin_asym_09b",
  any_mni152 = TRUE,
  surfaces = TRUE,
  as_job = TRUE
)

Arguments

subject

'RAVE' subject instance or character; see as_rave_subject

atlas_folders

paths to the atlas folders

template_name

template name where the atlases are created; see get_available_morph_to_template to get available templates and cmd_run_yael_preprocess to generate transforms to the templates

any_mni152

if the template is 'MNI152', then whether to check other template folders; default is true

surfaces

whether to generate surfaces; if true, then a 'GIfTI' file and a 'STL' file will be created for each atlas, if applicable. The 'GIfTI' file will be in right-anterior-superior 'RAS' coordinate system to comply with the 'NIfTI' standard, and 'STL' will be in left-posterior-superior 'LPS' system that can be imported to 'ITK'-based software such as 'BrainLab' or 'ANTs'.

as_job

whether to run as a background job for cleaner environment; default is true; set to false for debugging.

Value

The function returns nothing, but will create a folder named 'atlases' under raw subject 'rave-imaging' folder.

Examples

# Please check out https://rave.wiki to configure Python for RAVE
# or run ravemanager::configure_python()
if (FALSE) { # \dontrun{

generate_atlases_from_template("YAEL/yael_demo_001")

} # }