ravecore: Core File Structures and Workflows for 'RAVE'

Read, process, and analyze intracranial electroencephalography and deep-brain stimulation with 'RAVE', a reproducible framework for analysis and visualization of 'iEEG' by Magnotti, Wang, and Beauchamp, (2020, doi:10.1016/j.neuroimage.2020.117341 ). Supports 'BIDS' https://bids.neuroimaging.io or native file conventions to ingest signals in 'HDF5', 'Matlab', 'EDF', 'BrainVision', or 'BlackRock' formats; process images in 'NIfTI' and 'FreeSurfer' formats, providing brain imaging normalization to template brain, facilitate 'threeBrain' package for comprehensive electrode localization, see "your advanced electrode localizer" ('YAEL') by Wang, Magnotti, Zhang, and Beauchamp (2023, doi:10.1523/ENEURO.0328-23.2023 ).

See also

Author

Maintainer: Zhengjia Wang dipterix.wang@gmail.com (ORCID)

Authors:

John Magnotti john.magnotti@pennmedicine.upenn.edu

Xiang Zhang xiang.zhang@pennmedicine.upenn.edu

Michael Beauchamp michael.beauchamp@pennmedicine.upenn.edu

Other contributors:

Trustees of the University of Pennsylvania (All files in this package unless explicitly stated in the file) [copyright holder]