ravecore: Core File Structures and Workflows for 'RAVE'

Defines storage standard for Read, process, and analyze intracranial electroencephalography and deep-brain stimulation in 'RAVE', a reproducible framework for analysis and visualization of iEEG by Magnotti, Wang, and Beauchamp, (2020, doi:10.1016/j.neuroimage.2020.117341 ). Supports brain imaging data structure (BIDS) https://bids.neuroimaging.io and native file structure to ingest signals from 'Matlab' data files, hierarchical data format 5 (HDF5), European data format (EDF), BrainVision core data format (BVCDF), or BlackRock Microsystem (NEV/NSx); process images in Neuroimaging informatics technology initiative (NIfTI) and 'FreeSurfer' formats, providing brain imaging normalization to template brain, facilitating 'threeBrain' package for comprehensive electrode localization via 'YAEL' (your advanced electrode localizer) by Wang, Magnotti, Zhang, and Beauchamp (2023, doi:10.1523/ENEURO.0328-23.2023 ).

See also

Author

Maintainer: Zhengjia Wang dipterix.wang@gmail.com (ORCID)

Authors:

Xiang Zhang xiang.zhang@pennmedicine.upenn.edu

John Magnotti john.magnotti@pennmedicine.upenn.edu

Michael Beauchamp michael.beauchamp@pennmedicine.upenn.edu

Other contributors:

Trustees of the University of Pennsylvania (All files in this package unless explicitly stated in the file) [copyright holder]